NFDIV-755: Move post file to correct directory after merge with master
